Added logout and profile feature

+++ Envoy.blade.php
@@ -13,9 +13,23 @@
     start_containers
 @endstory
 
+@task('shut_down_existing_containers')
+    echo 'Shutting down existing deployment if exists'
+    if [ -d {{ $project_dir }} ]
+    then
+        echo 'Shutting down existing deployment'
+        cd {{ $project_dir }}
+        docker-compose -f docker-compose.staging.yml --env-file ./.env.staging down
+    fi
+@endtask
+
 @task('delete_existing_project')
-    echo 'Deleting existing project folder'
-    rm -rf {{ $project_dir }}
+    echo 'Deleting existing project folder if exists'
+    if [ -d {{ $project_dir }} ]
+    then
+        echo 'Deleting existing project folder'
+        rm -rf {{ $project_dir }}
+    fi
 @endtask
 
 @task('clone_repository')
@@ -25,39 +39,40 @@
     git reset --hard {{ $commit }}
 @endtask
 
+@task('create_iotportaldata_dir')
+    echo 'Deleting existing iotportaldata directory if exists'
+    if [ -d {{ $project_dir }} ]
+    then
+        echo 'Deleting existing iotportaldata directory'
+        rm -rf iotportaldata
+        echo 'Creating new iotportaldata directories'
+        mkdir -p iotportaldata/app/storage/app
+        mkdir -p iotportaldata/app/storage/framework
+        mkdir -p iotportaldata/ca-certificates
+        mkdir -p iotportaldata/env
+        mkdir -p iotportaldata/logs/app
+        mkdir -p iotportaldata/logs/nginx
+        mkdir -p iotportaldata/mysql/data
+        mkdir -p iotportaldata/nginx/sites-available
+        mkdir -p iotportaldata/ssl
+    fi
+@endtask
+
 @task('build_artifacts')
     echo 'Building artifacts'
     cd {{ $project_dir }}
     sed -i 's~server_name localhost host.docker.internal~server_name {{ $serverName }}~g' docker-compose/nginx/sites/default.conf
-{{--    sed -i 's~allow 127.0.0.1~allow {{ $unblockIp }}~g' docker-compose/nginx/sites/default.conf--}}
-{{--    sed -i 's~allow 127.0.0.2~allow {{ $serverIp }}~g' docker-compose/nginx/sites/default.conf--}}
     sed -i 's~APP_URL=.*~APP_URL={{ $appUrl }}~g' .env.staging
     sed -i 's~DB_PASSWORD=.*~DB_PASSWORD={{ $dbPassword }}~g' .env.staging
     sed -i 's~MQTT_HOST=.*~MQTT_HOST={{ $mqttHost }}~g' .env.staging
     sed -i 's~VMQ_WEBHOOKS_AUTH_ON_REGISTER_ENDPOINT=.*~VMQ_WEBHOOKS_AUTH_ON_REGISTER_ENDPOINT="${MIX_API_ENDPOINT}/mqtt/endpoint"~g' .env.staging
     sed -i 's~VMQ_WEBHOOKS_AUTH_ON_SUBSCRIBE_ENDPOINT=.*~VMQ_WEBHOOKS_AUTH_ON_SUBSCRIBE_ENDPOINT="${MIX_API_ENDPOINT}/mqtt/endpoint"~g' .env.staging
     sed -i 's~VMQ_WEBHOOKS_AUTH_ON_PUBLISH_ENDPOINT=.*~VMQ_WEBHOOKS_AUTH_ON_PUBLISH_ENDPOINT="${MIX_API_ENDPOINT}/mqtt/endpoint"~g' .env.staging
-    docker build -t inteliotportal-build -f docker-compose/build/Dockerfile .
-    #cp .env.staging .env
-    #echo '.env Content'
-    #head -10 .env
-    #echo 'Deleting old JS files'
-    #rm public/js/*.js
-    #rm public/js/*.txt
-    #echo 'Downloading NodeJS'
-    #curl -fsSL https://deb.nodesource.com/setup_14.x | sudo -E bash -
-    #sudo apt-get install -y nodejs
-    #echo 'Building JS files'
-    #npm ci
-    #npm rebuild node-sass
-    #npm run production
+    docker build --no-cache -t inteliotportal-build -f docker-compose/build/Dockerfile .
+    docker run -it -rm --name setup -v ../iotportaldata:/iotportaldata  inteliotportal-build
 @endtask
 
-@task('shut_down_existing_containers')
-    echo "Shutting down existing deployment"
-    cd {{ $project_dir }}
-    docker-compose -f docker-compose.staging.yml --env-file ./.env.staging down
-@endtask
+
 
 @task('start_containers')
     echo "Starting deployment ({{ $release }})"
@@ -81,6 +96,96 @@
 
 
 
+
+
+
+
+{{--Version 2--}}
+
+{{--@servers(['staging' => $stagingSshHost])--}}
+
+{{--@setup--}}
+{{--    $project_dir = 'i-iot-portal';--}}
+{{--    $release = date('YmdHis');--}}
+{{--@endsetup--}}
+
+{{--@story('deploy')--}}
+{{--    delete_existing_project--}}
+{{--    clone_repository--}}
+{{--    build_artifacts--}}
+{{--    shut_down_existing_containers--}}
+{{--    start_containers--}}
+{{--@endstory--}}
+
+{{--@task('delete_existing_project')--}}
+{{--    echo 'Deleting existing project folder'--}}
+{{--    rm -rf {{ $project_dir }}--}}
+{{--@endtask--}}
+
+{{--@task('clone_repository')--}}
+{{--    echo 'Cloning repository'--}}
+{{--    git clone --depth 1 {{ $projectRepository }}--}}
+{{--    cd {{ $project_dir }}--}}
+{{--    git reset --hard {{ $commit }}--}}
+{{--@endtask--}}
+
+{{--@task('build_artifacts')--}}
+{{--    echo 'Building artifacts'--}}
+{{--    cd {{ $project_dir }}--}}
+{{--    sed -i 's~server_name localhost host.docker.internal~server_name {{ $serverName }}~g' docker-compose/nginx/sites/default.conf--}}
+{{--    #sed -i 's~allow 127.0.0.1~allow {{ $unblockIp }}~g' docker-compose/nginx/sites/default.conf--}}
+{{--    #sed -i 's~allow 127.0.0.2~allow {{ $serverIp }}~g' docker-compose/nginx/sites/default.conf--}}
+{{--    sed -i 's~APP_URL=.*~APP_URL={{ $appUrl }}~g' .env.staging--}}
+{{--    sed -i 's~DB_PASSWORD=.*~DB_PASSWORD={{ $dbPassword }}~g' .env.staging--}}
+{{--    sed -i 's~MQTT_HOST=.*~MQTT_HOST={{ $mqttHost }}~g' .env.staging--}}
+{{--    sed -i 's~VMQ_WEBHOOKS_AUTH_ON_REGISTER_ENDPOINT=.*~VMQ_WEBHOOKS_AUTH_ON_REGISTER_ENDPOINT="${MIX_API_ENDPOINT}/mqtt/endpoint"~g' .env.staging--}}
+{{--    sed -i 's~VMQ_WEBHOOKS_AUTH_ON_SUBSCRIBE_ENDPOINT=.*~VMQ_WEBHOOKS_AUTH_ON_SUBSCRIBE_ENDPOINT="${MIX_API_ENDPOINT}/mqtt/endpoint"~g' .env.staging--}}
+{{--    sed -i 's~VMQ_WEBHOOKS_AUTH_ON_PUBLISH_ENDPOINT=.*~VMQ_WEBHOOKS_AUTH_ON_PUBLISH_ENDPOINT="${MIX_API_ENDPOINT}/mqtt/endpoint"~g' .env.staging--}}
+{{--    docker build -t inteliotportal-build -f docker-compose/build/Dockerfile .--}}
+{{--    #cp .env.staging .env--}}
+{{--    #echo '.env Content'--}}
+{{--    #head -10 .env--}}
+{{--    #echo 'Deleting old JS files'--}}
+{{--    #rm public/js/*.js--}}
+{{--    #rm public/js/*.txt--}}
+{{--    #echo 'Downloading NodeJS'--}}
+{{--    #curl -fsSL https://deb.nodesource.com/setup_14.x | sudo -E bash ---}}
+{{--    #sudo apt-get install -y nodejs--}}
+{{--    #echo 'Building JS files'--}}
+{{--    #npm ci--}}
+{{--    #npm rebuild node-sass--}}
+{{--    #npm run production--}}
+{{--@endtask--}}
+
+{{--@task('shut_down_existing_containers')--}}
+{{--    echo "Shutting down existing deployment"--}}
+{{--    cd {{ $project_dir }}--}}
+{{--    docker-compose -f docker-compose.staging.yml --env-file ./.env.staging down--}}
+{{--@endtask--}}
+
+{{--@task('start_containers')--}}
+{{--    echo "Starting deployment ({{ $release }})"--}}
+{{--    cd {{ $project_dir }}--}}
+{{--    docker-compose -f docker-compose.staging.yml --env-file ./.env.staging up -d --force-recreate --build--}}
+{{--@endtask--}}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+{{--Version 1--}}
 
 {{--@servers(['web' => 'root@174.138.35.53'])--}}
 {{----}}
